#1 Test moved to demo project because it's depends on clickhouse connection and build was failed.
